requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing & Export
pandas>=2.1.0
openpyxl>=3.1.0

# CLI enhancements
colorama>=0.4.6
tqdm>=4.66.0
